// src/models/Metrics.ts
export enum DailyMetric {
  DAILY_METRIC_UNKNOWN = 'DAILY_METRIC_UNKNOWN',
  BUSINESS_IMPRESSIONS_DESKTOP_MAPS = 'BUSINESS_IMPRESSIONS_DESKTOP_MAPS',
  BUSINESS_IMPRESSIONS_DESKTOP_SEARCH = 'BUSINESS_IMPRESSIONS_DESKTOP_SEARCH',
  BUSINESS_IMPRESSIONS_MOBILE_MAPS = 'BUSINESS_IMPRESSIONS_MOBILE_MAPS',
  BUSINESS_IMPRESSIONS_MOBILE_SEARCH = 'BUSINESS_IMPRESSIONS_MOBILE_SEARCH',
  BUSINESS_CONVERSATIONS = 'BUSINESS_CONVERSATIONS',
  BUSINESS_DIRECTION_REQUESTS = 'BUSINESS_DIRECTION_REQUESTS',
  CALL_CLICKS = 'CALL_CLICKS',
  WEBSITE_CLICKS = 'WEBSITE_CLICKS',
  BUSINESS_BOOKINGS = 'BUSINESS_BOOKINGS',
  BUSINESS_FOOD_ORDERS = 'BUSINESS_FOOD_ORDERS',
  BUSINESS_FOOD_MENU_CLICKS = 'BUSINESS_FOOD_MENU_CLICKS',
}

export interface DateRange {
  year: number;
  month: number;
  day: number;
}

export interface DailyRange {
  start_date: DateRange;
  end_date: DateRange;
}

export interface PerformanceMetricsRequest {
  dailyMetrics: DailyMetric[];
  dailyRange: DailyRange;
}